+++ pkg/webui/console/containers/freq-plans-select/utils.js
@@ -14,12 +14,13 @@
 
 import { defineMessages } from 'react-intl'
 
-export const formatOptions = plans => plans.map(plan => ({ value: plan.id, label: plan.name }))
+export const formatOptions = plans =>
+  plans.map(plan => ({ value: plan.id, label: plan.name, tag: plan.band_id }))
 export const m = defineMessages({
   warning: 'Frequency plans unavailable',
   none: 'Do not set a frequency plan',
   selectFrequencyPlan: 'Select a frequency plan...',
   addFrequencyPlan: 'Add frequency plan',
   frequencyPlanDescription:
-    'Note: most gateways use a single frequency plan. Some 16 and 64 channel gateways however allow setting multiple.',
+    'Note: most gateways use a single frequency plan. Some 16 and 64 channel gateways however allow setting multiple within the same band.',
 })